+++ knowage-core/src/main/java/it/eng/spagobi/sdk/utilities/SDKObjectsConverter.java
@@ -34,6 +34,8 @@ import org.apache.axis.attachments.ManagedMemoryDataSource;
 import org.apache.log4j.Logger;
 import org.json.JSONObject;
 
+import it.eng.knowage.commons.security.PathTraversalChecker;
+import it.eng.knowage.commons.security.exceptions.PathTraversalAttackException;
 import it.eng.qbe.dataset.QbeDataSet;
 import it.eng.spago.base.SourceBean;
 import it.eng.spago.base.SourceBeanException;
@@ -228,8 +230,15 @@ public class SDKObjectsConverter {
 			}
 			if (dh != null && dh.getName() != null) {
 				LOGGER.debug("Deleting attachment file ...");
-				File attachment = new File(dh.getName());
-				if (attachment.exists() && attachment.isFile()) {
+				String fileName = dh.getName();
+				File attachment = null;
+				try {
+					PathTraversalChecker.isValidFileName(fileName);
+					attachment = new File(fileName);
+				} catch (Exception e) {
+					throw new PathTraversalAttackException("Error converting SDKTemplate " + sdkTemplate + ". File name " + fileName + " is not valid.");
+				}
+				if (attachment != null && attachment.exists() && attachment.isFile()) {
 					boolean attachmentFileDeleted;
 					try {
 						attachmentFileDeleted = Files.deleteIfExists(attachment.toPath());
